adds base mirage handlers
