+++ ui/mirage/handlers/index.js
@@ -1,5 +1,5 @@
 // add all handlers here
 // individual lookup done in mirage config
-import example from './example';
+import base from './base';
 
-export { example };
+export { base };
